Team Yorch changes

+++ CSH/library/api/fermat-csh-api/src/main/java/com/bitdubai/fermat_csh_api/layer/csh_wallet/cash_money/interfaces/CashMoneyBalanceRecord.java
@@ -24,7 +24,7 @@ public interface CashMoneyBalanceRecord {
 
     TransactionType getTransactionType();
 
-    float getAmount();
+    double getAmount();
 
     CashCurrencyType getCashCurrencyType();
